+++ e2e/lib/e2e_web/controllers/page_controller.ex
@@ -1117,6 +1117,10 @@ defmodule E2eWeb.PageController do
     render(conn, :data_list_page)
   end
 
+  def data_list_styling_page(conn, _params) do
+    render(conn, :data_list_styling_page)
+  end
+
   def data_table_page(conn, _params) do
     render(conn, :data_table_page)
   end