+++ e2e/lib/e2e_web/demos/radio_group_demo.ex
@@ -1474,6 +1474,4 @@ defmodule E2eWeb.Demos.RadioGroupDemo do
     </.radio_group>
     """
   end
-
-
 end