+++ e2e/lib/e2e_web/live/pages_live/select_patterns_live.ex
@@ -97,7 +97,12 @@ defmodule E2eWeb.SelectPatternsLive do
           title="Stream"
           code_tabs={[
             %{value: "heex", label: "Heex", language: :heex, code: Demo.patterns_stream_demo_heex()},
-            %{value: "elixir", label: "Elixir", language: :elixir, code: Demo.patterns_stream_elixir()}
+            %{
+              value: "elixir",
+              label: "Elixir",
+              language: :elixir,
+              code: Demo.patterns_stream_elixir()
+            }
           ]}
         >
           <:preview>